requirements-cpu.txt
solana==0.29.1
solders>=0.14.0,<0.15.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.24.0,<2.0.0
tenacity>=8.2.3
cryptography>=41.0.7
cachetools>=4.2.2,<5.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.23
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
base58==2.1.1
google-generativeai>=0.5.4
gensim>=4.3.0
shap>=0.41.0
lime>=0.2.0
PyPortfolioOpt>=1.5.0
Riskfolio-Lib>=4.0.0
python-socketio[fastapi]>=5.8.0
fastapi-limiter>=0.1.5
aiohttp>=3.9.1
async-timeout>=4.0.3
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
httpx>=0.23.0,<0.24.0
psutil==5.9.6 